move pkg/health out of registry-scanner
